Move email to user actions
